aio_read() aio_write() return int, not ssize_t
compilation used to fail on 64bit machines.

+++ replayer/Commander.cpp
@@ -211,7 +211,7 @@ void AsynchronousCommander::execute(uint64_t operation,
 
 	control_block_queue.push(op);
 
-	ssize_t (*aio_op)(aiocb64*) = (operation == OPERATION_READ) ?
+	int (*aio_op)(aiocb64*) = (operation == OPERATION_READ) ?
 			aio_read64 : aio_write64;
 
 	stats->currentDelay = waitUntil(time);
